+++ source/extensions/isaacsim.ros2.nodes/premake5.lua
@@ -40,7 +40,6 @@ includedirs {
     "%{kit_sdk_bin_dir}/dev/fabric/include/",
     "%{root}/_build/target-deps/nlohmann_json/include",
     "%{root}/source/extensions/isaacsim.core.includes/include",
-    "%{root}/source/extensions/isaacsim.core.nodes/include",
     "%{root}/_build/target-deps/cuda/include",
     "%{root}/_build/target-deps/python/include/python3.11",
     "%{root}/_build/target-deps/python/include",